Register/Login screen simple functions + UI refactoring

+++ app/src/main/java/com/assist_software/expenseappmvp/data/database/entities/User.kt
@@ -6,9 +6,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "Users")
 data class User (
-        @PrimaryKey(autoGenerate = true) var userId: Int = 0,
+        @PrimaryKey(autoGenerate = true) var userId: Long = 0,
         @ColumnInfo(name = "userName") var userName: String = "",
         @ColumnInfo(name = "userEmail") var userEmail: String = "",
         @ColumnInfo(name = "userPassword") var userPassword: String = "",
-        @ColumnInfo(name = "userCurrentBalance") var userCurrentBalance: Long = 0
+        @ColumnInfo(name = "userCurrentBalance") var userCurrentBalance: Double = 0.0
 )
